toggle-feature-v1
toggle feature working in this version, had to make modifications to the discordApp file because I didn't register the toggle as a command. The code is just looking for an instance where a member sends the message "!toggle" and then seeing if they have the permissions to toggle the bot.